Phase 2 demonstration
Test_Script_2.cs is an temporary script which performs atomic operations like highlighting the player and moving player from point A to B only. The desired operation is however a state machine that performs intermediate operations in-between.
Dictionary keywords.Add() throws out of bound exception. So we must use a foreach loop.